+++ sjf4j/src/main/java/org/sjf4j/annotation/compiled/PutIfParentPresentByPath.java
@@ -1,0 +1,12 @@
+package org.sjf4j.annotation.compiled;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.METHOD)
+@Retention(RetentionPolicy.SOURCE)
+public @interface PutIfParentPresentByPath {
+    String value();
+}